refactor(video): 重构视频信息结构体字段类型和枚举定义
- 将视频合集作者id字段从String类型改为u64类型
- 将视频合集中分部id字段从u32类型改为Option<u32>类型
- 将视频工作人员会员类型从u8改为VipType枚举类型
- 将视频工作人员会员状态从u8改为bool类型并添加序列化转换
- 将视频工作人员认证级别从u8改为OfficialRole枚举类型
- 将视频工作人员认证类型从i8改为bool类型并添加序列化转换
- 新增VideoHonorType枚举类型替换原有的u8占位符
- 移除调试用的println语句
- 在build-builder中添加文档注释支持
- 将布尔值参数转换为字符串时统一使用u8强制转换
- 新增VipType枚举定义用于表示会员类型
- 从 Cargo.toml 中移除 proc-macro2 依赖
- 从 Cargo.lock 中移除 build-builder 对 proc-macro2 的依赖记录
- 保持 quote 依赖的正常引用
- 优化构建配置文件的依赖管理

+++ src/video/stream.rs
@@ -362,7 +362,7 @@ impl VideoStreamArgs {
             ("qn", self.qn.map(|q| u32::from(&q).to_string())),
             ("fnval", Some(u32::from(&self.fanval).to_string())),
             ("fnver", Some(self.fnver.to_string())),
-            ("fourk", Some(self.fourk.to_string())),
+            ("fourk", Some((self.fourk as u8).to_string())),
             ("session", self.session.clone()),
             ("otype", self.otype.clone()),
             ("type", self.r#type.clone()),
@@ -370,8 +370,8 @@ impl VideoStreamArgs {
                 "platform",
                 Some(<&VideoPlatform as Into<&str>>::into(&self.platform).to_string()),
             ),
-            ("high_quality", Some(self.high_quality.to_string())),
-            ("try_look", Some(self.try_look.to_string())),
+            ("high_quality", Some((self.high_quality as u8).to_string())),
+            ("try_look", Some((self.try_look as u8).to_string())),
         ]
     }
 }
